Account Managment to reactive

+++ user-management/src/main/java/org/shaft/administration/usermanagement/entity/User.java
@@ -10,7 +10,7 @@ import org.springframework.data.elasticsearch.annotations.Document;
 public class User {
     @Id
     int i;
-    int c;
+    long c;
     int a;
     String nm;
     String e;
